Restaurant revenue + Order total price

+++ Delivery.Api.DAL.Common/Entities/DishEntity.cs
@@ -9,14 +9,14 @@ namespace Delivery.Api.DAL.Common.Entities
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public Decimal Price { get; set; }
         public Guid RestaurantId { get; set; }
         public RestaurantEntity? Restaurant { get; set; }
         public string? ImageUrl { get; set; }
         // public ICollection<Allergen> Allergens { get; set; } = new List<Allergen>();
         public Collection<DishAmountEntity> DishAmounts { get; set; } = new Collection<DishAmountEntity>();
 
-        public DishEntity(Guid id, string name, string description, double price, Guid restaurantId, string? imageUrl = null) : base(id)
+        public DishEntity(Guid id, string name, string description, Decimal price, Guid restaurantId, string? imageUrl = null) : base(id)
         {
             Name = name;
             Description = description;
